Change inheritance and association mapping
The @MappedSuperclass one-to-many with anoter @MappedSuperclass
was not possible. This new solution uses bidirectional associations
with one @MappedSuperclass and another single table inheritance.

Now it's possible to have Page and Container implementing a Collection
and also Text and Form implementing an Item and provide a one-to-many
relation between a Collection and multiple Items.

Perhaps we should change in the future to unidirectional many-to-one,
making it easier to implement new Item classes witout the need to update
the Collection superclass with new properties. That change would mean
the service class wouldn't fetch the Collection and then query all the
items, but rather query all the items belonging to a specific Collection.