Make service class a hard dependency
Because we need this service for all actions, it's better to mark
the article service as a hard dependency for the article and archive
controllers.

Also the blog id is now detached from the method where
the articles are queries, so the methods correspond better with the
actual work done in the method.