Change assocation collection<->item to unidirectional
The unidirectional relation makes it easier to relate items to a
specific collection. The service class is injected with all possible
items and queries each of them to find items for the current collection.

Other modules can easily add more items to the service class to enrich
the module with additional items.